--- CodeStarterCamp_Week1/Test.swift
@@ -1,28 +1,0 @@
-//
-//  main.swift
-//  CodeStarterCamp_Week1
-//
-//  Created by Yeji on 11/18/23.
-//
-
-import Foundation
-
-func test() {
-    func plusTwo(x: Int) -> Int {  // parameter
-        return x + 2 // 반환값
-    }
-
-    var a: Int
-    a = plusTwo(x: 4)
-    print(a)
-    
-    func printHelloWorld() {
-        print("Hello,World")
-    }
-    
-    printHelloWorld()
-}
-
-
-
-
