+++ playground/Loop.playground/Contents.swift
@@ -1,5 +1,36 @@
 import Swift
 
+
+for _ in 1...3 {
+    print("하이!")
+}
+
+for count in 1...3 {
+    print(count, "번째 무야호!!")
+}
+
+// 스스로 해보기 1
+func printMuYaHo(count: Int) {
+    print(count, "번째 무야호!!")
+}
+
+for count in 0...4 {
+    printMuYaHo(count: count)
+}
+
+// 스스로 해보기 2
+func gugu(x: Int) {
+    for count in 1...9 {
+        print("\(count)*\(x)=\(x*count)")
+    }
+}
+
+gugu(x: 3)
+gugu(x: 8)
+
+
+/* 시작 */
+
 var integers = [1, 2, 3]
 let people = ["yagom": 10, "eric": 15, "mike": 12]
